refactoring test packages structure
