+++ src/test/java/juja/microservices/gamification/BaseIntegrationTest.java
@@ -1,4 +1,4 @@
-package juja.microservices.gamification.integration;
+package juja.microservices.gamification;
 
 import com.github.fakemongo.Fongo;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbConfiguration;
@@ -6,8 +6,6 @@ import com.lordofthejars.nosqlunit.mongodb.MongoDbRule;
 import com.lordofthejars.nosqlunit.mongodb.SpringMongoDbRule;
 import com.mongodb.MockMongoClient;
 import com.mongodb.MongoClient;
-import juja.microservices.gamification.Gamification;
-import juja.microservices.gamification.GamificationTestConfig;
 import org.junit.Rule;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.web.context.WebApplicationContext;
