+++ src/test/java/juja/microservices/gamification/controller/AchievementValidationJsonIntegrationTest.java
@@ -1,6 +1,7 @@
-package juja.microservices.gamification.integration;
+package juja.microservices.gamification.controller;
 
 import com.lordofthejars.nosqlunit.annotation.UsingDataSet;
+import juja.microservices.gamification.BaseIntegrationTest;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
